+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Add SavedSearches menu on MapExplorer and SavedSearchesTab in UserSettings. Users can view a list of their saved searches and delete a saved search from either of these locations. Users can create a saved search from the MapExplorer view.
+
+### Changed
+- Notification opt-in prompt covers saved searches too (in addition to communities)
 - Fix editing of values on user sign-up and community create review screens (fixes #685)
 - Remove editing of user name and email from Community Creation flow (fixes #579, #582)